refactor(write_engine): split mod.rs into maintenance/compaction/stats submodules (#1120)
Behavior-preserving module split (epic #1116). Extract concerns out of the
4,395-line write_engine/mod.rs into sibling submodules, all reachable as
in-crate `impl WriteEngine` blocks (private fields stay private to the crate):

- maintenance.rs: STCS maintenance_step state machine, candidate scanning,
  startup orphan sweeps, atomic input deletion, MaintenanceReport + ActiveMerge.
- compaction.rs: merge execution (start_merge + blocking/async finalize) split
  out of maintenance to keep each file under the size ceiling.
- stats.rs: lightweight engine counters (memtable/wal sizes, total_written,
  l0_count) + CompactionStats snapshot type.
- test_support.rs: shared #[cfg(test)] fixtures (create_test_schema,
  create_test_mutation, flush_n_sstables_sync).

mod.rs keeps config types, the WriteEngine struct, lifecycle (new/write/
flush/close/execute), generation helpers, Drop, and re-exports the moved
public types (MaintenanceReport, CompactionStats) so the public surface is
unchanged. Method bodies moved verbatim; the only edits are pub(crate)
visibility on cross-module-called fns/fields. Test count preserved (62).

mod.rs 4,395 -> 2,070 lines.
