+++ cqlite-core/src/storage/write_engine/mod.rs
@@ -53,51 +53,28 @@ pub use mutation::{
 #[cfg(feature = "write-support")]
 pub use wal::WriteAheadLog;
 
+#[cfg(feature = "write-support")]
+mod compaction;
+#[cfg(feature = "write-support")]
+mod maintenance;
+#[cfg(feature = "write-support")]
+mod stats;
+#[cfg(all(test, feature = "write-support"))]
+mod test_support;
+
+#[cfg(feature = "write-support")]
+pub use maintenance::MaintenanceReport;
+#[cfg(feature = "write-support")]
+pub use stats::CompactionStats;
+
 use crate::error::{Error, Result};
 use crate::schema::{TableSchema, UdtRegistry};
 use crate::storage::sstable::writer::SSTableInfo;
+#[cfg(feature = "write-support")]
+use maintenance::ActiveMerge;
 use std::path::{Path, PathBuf};
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::time::{Duration, Instant};
-
-/// Maintenance report from a maintenance_step() call (M5.2, Issue #384)
-#[cfg(feature = "write-support")]
-#[derive(Debug, Clone)]
-pub struct MaintenanceReport {
-    /// Time spent in this maintenance step
-    pub time_spent: Duration,
-    /// Completed merge output files (if any merge completed)
-    pub completed_merges: Vec<PathBuf>,
-    /// Number of rows merged in this step
-    pub rows_merged: u64,
-    /// Number of bytes written in this step
-    pub bytes_written: u64,
-    /// Whether there is pending compaction work
-    pub pending_compaction: bool,
-}
-
-/// Cumulative statistics across all compaction operations (M5.2, Issue #474)
-///
-/// Tracks lifetime totals for monitoring compaction health and throughput.
-/// Updated atomically at the end of each successful merge.
-#[cfg(feature = "write-support")]
-#[derive(Debug, Clone, Default)]
-pub struct CompactionStats {
-    /// Total number of completed compaction cycles
-    pub compactions_completed: u64,
-    /// Total number of input SSTables consumed
-    pub sstables_merged_in: u64,
-    /// Total number of output SSTables produced
-    pub sstables_produced: u64,
-    /// Total bytes read from input SSTables
-    pub bytes_read: u64,
-    /// Total bytes written to output SSTables
-    pub bytes_written: u64,
-    /// Total rows merged across all compactions
-    pub rows_merged: u64,
-    /// Total wall-clock time spent in compaction
-    pub total_time: Duration,
-}
+use std::time::Instant;
 
 /// Trait for merge policy implementations (M5.2, Issue #383)
 ///
@@ -117,45 +94,6 @@ pub trait MergePolicy: Send + std::fmt::Debug {
     /// Paths to SSTables that should be merged, ordered newest to oldest.
     /// Returns empty Vec if no compaction is needed.
     fn select_merge(&self, candidates: &[PathBuf]) -> Result<Vec<PathBuf>>;
-}
-
-/// Active merge state for incremental compaction (M5.2, Issue #384)
-#[cfg(feature = "write-support")]
-#[derive(Debug)]
-struct ActiveMerge {
-    /// K-way merger performing the compaction
-    merger: KWayMerger,
-    /// Output SSTable writer (writes to `tmp_dir/keyspace/table/`)
-    writer: crate::storage::sstable::writer::SSTableWriter,
-    /// Input SSTable paths being merged (these remain intact until atomic rename succeeds)
-    input_paths: Vec<PathBuf>,
-    /// Root of the temporary directory tree used for this compaction output.
-    ///
-    /// The SSTableWriter appends `keyspace/table/` to this path, so component
-    /// files land at `tmp_dir/keyspace/table/nb-{gen}-big-*.{ext}`.
-    ///
-    /// After `writer.finish()` the files are atomically renamed to the final
-    /// SSTable directory. Only then are the inputs deleted.
-    ///
-    /// Invariant: if the process crashes before the renames complete, `tmp_dir`
-    /// may contain partial output but the input SSTables remain intact.
-    tmp_dir: PathBuf,
-    /// Final SSTable directory (`data_dir/keyspace/table/`)
-    ///
-    /// Stored here so `finalize_merge_async` doesn't have to recompute it.
-    sstable_dir: PathBuf,
-    /// Number of rows merged so far (updated per partition)
-    rows_merged: u64,
-    /// Total bytes read from input SSTables (approximate: sum of Data.db file sizes)
-    bytes_read: u64,
-    /// When this merge started
-    started_at: Instant,
-    /// Effective compaction schema (#850): the configured schema augmented with
-    /// any static columns that appear in the input SSTables' SerializationHeaders
-    /// but were dropped from the current schema. Used to convert merged entries to
-    /// mutations so the writer still emits the static-row prelude (static-column
-    /// presence is read from the input headers, not the current schema only).
-    effective_schema: TableSchema,
 }
 
 /// WAL durability mode for the write engine.
@@ -1002,51 +940,6 @@ impl WriteEngine {
         Ok(())
     }
 
-    /// Get the current memtable size in bytes
-    pub fn memtable_size(&self) -> usize {
-        self.memtable.size_bytes()
-    }
-
-    /// Get the current memtable row count
-    pub fn memtable_row_count(&self) -> usize {
-        self.memtable.row_count()
-    }
-
-    /// Get the current WAL size in bytes
-    pub fn wal_size(&self) -> u64 {
-        self.wal.size()
-    }
-
-    /// Get the current generation number
-    pub fn generation(&self) -> u64 {
-        self.generation
-    }
-
-    /// Return the cumulative number of rows written since the engine was opened
-    /// (Issue #486).
-    ///
-    /// This counter is incremented for every row that is successfully inserted
-    /// into the memtable and is NOT reset on flush.  It therefore represents
-    /// the total write throughput for the current session.
-    ///
-    /// Note: This counter is in-process only and resets to zero when the engine
-    /// is re-opened.  WAL replay rows (recovered from a previous crash) are NOT
-    /// counted; only rows written through `write()` / `write_async()` during
-    /// the current session are counted.
-    pub fn total_written(&self) -> u64 {
-        self.rows_written
-    }
-
-    /// Return the number of L0 SSTables successfully flushed since the engine
-    /// was opened (Issue #486).
-    ///
-    /// Incremented once per successful `flush()` call that produces a non-empty
-    /// SSTable.  This is an in-process counter and resets to zero when the
-    /// engine is re-opened.
-    pub fn l0_count(&self) -> u64 {
-        self.l0_count
-    }
-
     /// Parse a CQL statement to a Mutation
     ///
     /// Supports INSERT, UPDATE, and DELETE statements.
@@ -1105,1054 +998,6 @@ impl WriteEngine {
         }
         Ok(())
     }
-
-    /// Set the merge policy for background compaction (M5.2, Issue #383)
-    ///
-    /// # Arguments
-    ///
-    /// * `policy` - Merge policy implementation (e.g., STCS, LCS, TWCS)
-    pub fn set_merge_policy(&mut self, policy: Box<dyn MergePolicy>) -> Result<()> {
-        self.merge_policy = Some(policy);
-        Ok(())
-    }
-
-    /// Return cumulative compaction statistics (M5.2, Issue #474)
-    ///
-    /// Returns a snapshot of the lifetime totals accumulated across all compaction
-    /// cycles that have completed since the `WriteEngine` was created. The snapshot
-    /// is cheaply cloneable and safe to inspect from any thread (no lock required,
-    /// because `WriteEngine` itself is not `Sync`).
-    ///
-    /// # Example
-    ///
-    /// ```rust,ignore
-    /// let stats = engine.maintenance_stats();
-    /// println!(
-    ///     "Completed {} compactions, merged {} rows, wrote {} bytes",
-    ///     stats.compactions_completed,
-    ///     stats.rows_merged,
-    ///     stats.bytes_written,
-    /// );
-    /// ```
-    pub fn maintenance_stats(&self) -> CompactionStats {
-        self.cumulative_stats.clone()
-    }
-
-    /// Perform incremental maintenance work (M5.2, Issue #384)
-    ///
-    /// This method performs background compaction work within a time budget.
-    /// It can be called repeatedly from a background thread or task scheduler
-    /// to make incremental progress on compaction.
-    ///
-    /// ## Runtime contexts
-    ///
-    /// This is a synchronous method, but its internal async-to-sync bridge is
-    /// runtime-aware (see [`merge::block_on_async`]), so it is safe to call from
-    /// **either** a plain synchronous context **or** from within an active Tokio
-    /// runtime — including `#[tokio::main]`/`#[tokio::test]` worker threads and
-    /// `async fn` callers. Prior to Issue #587 calling it from inside a runtime
-    /// panicked with "Cannot start a runtime from within a runtime" once a merge
-    /// had input SSTables to read. The sync signature is preserved so the CLI and
-    /// Python bindings can keep calling it directly. (The Node binding wraps it in
-    /// `spawn_blocking`, which remains correct.)
-    ///
-    /// ## Behavior
-    ///
-    /// 1. If no active merge exists, consult the merge policy for work
-    /// 2. If merge work is available, start a new merge
-    /// 3. Process the active merge until budget is exhausted
-    /// 4. Return progress report
-    ///
-    /// ## Invariants
-    ///
-    /// - Budget is honored within 10% tolerance
-    /// - At least one partition is processed per call (minimum progress guarantee)
-    /// - Merge state is preserved across calls for resumption
-    ///
-    /// ## Budget Enforcement
-    ///
-    /// The budget is honored within approximately 10% tolerance. This tolerance
-    /// exists to avoid interrupting partition processing mid-stream, which would
-    /// require complex state management to resume. The tolerance ensures forward
-    /// progress on each call while remaining responsive to time constraints.
-    ///
-    /// # Arguments
-    ///
-    /// * `budget` - Maximum time to spend in this call
-    ///
-    /// # Returns
-    ///
-    /// A report containing progress metrics and whether more work is pending.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    /// - Engine has been closed
-    /// - Merge policy returns an error
-    /// - SSTable reading or writing fails
-    ///
-    /// # Example
-    ///
-    /// ```rust,ignore
-    /// use std::time::Duration;
-    ///
-    /// // Background compaction loop
-    /// loop {
-    ///     let report = engine.maintenance_step(Duration::from_millis(100))?;
-    ///
-    ///     if !report.pending_compaction {
-    ///         // No more work, sleep or exit
-    ///         break;
-    ///     }
-    ///
-    ///     // Log progress
-    ///     println!("Merged {} rows in {:?}", report.rows_merged, report.time_spent);
-    /// }
-    /// ```
-    #[tracing::instrument(name = "compaction.maintenance_step", skip(self))]
-    pub fn maintenance_step(&mut self, budget: Duration) -> Result<MaintenanceReport> {
-        // Budget requested for this step (issue #1037). Compared with the
-        // consumed budget below (the scheduler honors a ~10% tolerance).
-        crate::observability::record_histogram(
-            crate::observability::catalog::COMPACTION_BUDGET_REQUESTED,
-            budget.as_secs_f64(),
-            &[],
-        );
-
-        let result = self.maintenance_step_inner(budget);
-
-        // Budget consumed + lifetime-throughput counters (issue #1037). Recorded
-        // for every step (even a no-op one) so the budget-tolerance signal is
-        // complete; rows-merged is per-step and feeds the throughput rate when
-        // combined with COMPACTION_DURATION at finalize.
-        if let Ok(report) = &result {
-            use crate::observability::{self as obs, catalog};
-            obs::record_histogram(
-                catalog::COMPACTION_BUDGET_CONSUMED,
-                report.time_spent.as_secs_f64(),
-                &[],
-            );
-            obs::add_counter(catalog::COMPACTION_ROWS_MERGED, report.rows_merged, &[]);
-            obs::record_gauge(catalog::COMPACTION_LAG, self.l0_count as i64, &[]);
-        }
-
-        crate::observability::record_result("compaction", result)
-    }
-
-    fn maintenance_step_inner(&mut self, budget: Duration) -> Result<MaintenanceReport> {
-        if self.closed.load(Ordering::SeqCst) {
-            return Err(Error::InvalidInput(
-                "WriteEngine has been closed".to_string(),
-            ));
-        }
-
-        let start = Instant::now();
-        let mut report = MaintenanceReport {
-            time_spent: Duration::from_secs(0),
-            completed_merges: Vec::new(),
-            rows_merged: 0,
-            bytes_written: 0,
-            pending_compaction: false,
-        };
-
-        // If no merge policy is set, no maintenance work to do
-        let merge_policy = match &self.merge_policy {
-            Some(policy) => policy,
-            None => {
-                report.time_spent = start.elapsed();
-                return Ok(report);
-            }
-        };
-
-        // If no active merge exists, check if we should start one
-        if self.active_merge.is_none() {
-            // SCOPE TO THIS TABLE (#935 branch review): `scan_sstable_candidates`
-            // walks the whole `data_dir` recursively, so it can include SSTables
-            // of OTHER keyspaces/tables. This WriteEngine is single-table
-            // (`config.schema`) and always publishes output to
-            // `data_dir/keyspace/table/`, so restrict the candidate set to THIS
-            // table's directory BEFORE any policy or purge-safety decision.
-            // Otherwise a full compaction of this table is misclassified as
-            // partial whenever a foreign table's SSTable exists under `data_dir`
-            // (`selected_set != candidate_set`), which both lets the policy see
-            // foreign-table inputs and disables tombstone purging that is actually
-            // safe. Every published SSTable for this table lives under
-            // `table_dir`, so the scoping never drops a real input.
-            let table_dir = self
-                .config
-                .data_dir
-                .join(&self.config.schema.keyspace)
-                .join(&self.config.schema.table);
-            let candidates: Vec<PathBuf> = self
-                .scan_sstable_candidates()?
-                .into_iter()
-                .filter(|p| p.starts_with(&table_dir))
-                .collect();
-            let selected = merge_policy.select_merge(&candidates)?;
-
-            if !selected.is_empty() {
-                // Overlap-safety gate for tombstone purging (#921 finding 1): a
-                // compaction may purge tombstones ONLY when it spans EVERY
-                // candidate SSTable for the table (a major/full compaction).
-                // Otherwise a tombstone could be purged while a non-included
-                // overlapping SSTable still holds data it shadows, resurrecting
-                // that data on the next read. A partial selection (the common
-                // background-compaction case) is therefore purge-UNSAFE: it
-                // retains tombstones. Compare as sets so input ordering does not
-                // affect the decision.
-                let selected_set: std::collections::HashSet<&PathBuf> = selected.iter().collect();
-                let candidate_set: std::collections::HashSet<&PathBuf> =
-                    candidates.iter().collect();
-                let purge_safe = !candidate_set.is_empty() && selected_set == candidate_set;
-
-                // Overlap-aware partial-compaction purging (#935): when this is a
-                // PARTIAL compaction (some candidate SSTables are NOT included),
-                // compute the min write timestamp across those non-included
-                // SSTables. A tombstone older than every one of them shadows
-                // nothing outside the set and can be purged even here. For a full
-                // compaction (`purge_safe == true`) there are no non-included
-                // SSTables, so the bound is `None` and the merger uses its +inf
-                // full-compaction fast path. `candidates` is already scoped to
-                // this table's directory (see above), so the non-included set is
-                // exactly this table's outside SSTables.
-                let max_purgeable_timestamp = if purge_safe {
-                    None
-                } else {
-                    let non_included: Vec<PathBuf> = candidates
-                        .iter()
-                        .filter(|p| !selected_set.contains(*p))
-                        .cloned()
-                        .collect();
-                    merge::compute_max_purgeable_timestamp(&non_included)
-                };
-
-                // Start a new merge
-                self.start_merge(selected, purge_safe, max_purgeable_timestamp)?;
-            } else {
-                // No work selected by policy
-                report.time_spent = start.elapsed();
-                report.pending_compaction = false;
-                return Ok(report);
-            }
-        }
-
-        // Process active merge within budget
-        let budget_tolerance = budget.mul_f32(1.1); // 10% tolerance
-        let mut partitions_processed = 0;
-
-        while let Some(merge) = &mut self.active_merge {
-            // Check budget (but always process at least one partition)
-            if partitions_processed > 0 && start.elapsed() >= budget_tolerance {
-                break;
-            }
-
-            // Process one partition from the merge
-            let step = merge.merger.step()?;
-
-            match step {
-                merge::MergeStep::Partition { key, rows } => {
-                    partitions_processed += 1;
-
-                    // Convert MergeEntry rows to Mutation format
-                    // (collect into a vec first to release the borrow on merge)
-                    let entries_vec: Vec<_> = rows.into_iter().collect();
-
-                    // Now we can call self methods without conflict.
-                    // Skip metadata-only entries (#886/#899 branch-review): they
-                    // carry complex/range deletion metadata through the merge
-                    // stream but have no writer-emittable content yet, so writing
-                    // them would produce a phantom live empty (pure-PK) row at
-                    // timestamp 0. See `MergeEntry::is_metadata_only_no_op`.
-                    // #850: convert with the effective compaction schema so any
-                    // static column re-added from the input headers is preserved
-                    // (partition-key decoding is identical; only static columns
-                    // differ). Falls back to the config schema if (impossibly) no
-                    // active merge is present.
-                    let conversion_schema = self
-                        .active_merge
-                        .as_ref()
-                        .map(|m| m.effective_schema.clone())
-                        .unwrap_or_else(|| self.config.schema.clone());
-                    let mutations = entries_vec
-                        .into_iter()
-                        .filter(|entry| !entry.is_metadata_only_no_op())
-                        .map(|entry| {
-                            merge::KWayMerger::merge_entry_to_mutation(entry, &conversion_schema)
-                        })
-                        .collect::<Result<Vec<_>>>()?;
-
-                    // If every merged row was metadata-only, the partition has no
-                    // writer-emittable content. Skip `write_partition` to avoid a
-                    // phantom EMPTY partition (header/end marker + Index/Filter/
-                    // Summary/statistics registration) in the output SSTable, and
-                    // do not count it as an output partition or row (#886
-                    // branch-review).
-                    if mutations.is_empty() {
-                        continue;
-                    }
-
-                    // Count rows actually written (skipped metadata-only entries
-                    // produce no row, so they must not inflate the stats). A pure
-                    // range-tombstone carrier (#933) or a pure partition-tombstone
-                    // carrier (#1072) emits a marker / partition-header deletion,
-                    // not a row — exclude them, matching KWayMerger::merge.
-                    let row_count = mutations
-                        .iter()
-                        .filter(|m| {
-                            let is_range_only = m.operations.is_empty()
-                                && m.partition_tombstone.is_none()
-                                && m.row_tombstone.is_none()
-                                && !m.range_tombstones.is_empty();
-                            let is_partition_only = m.operations.is_empty()
-                                && m.partition_tombstone.is_some()
-                                && m.row_tombstone.is_none()
-                                && m.range_tombstones.is_empty();
-                            !(is_range_only || is_partition_only)
-                        })
-                        .count() as u64;
-
-                    // Write partition to output SSTable
-                    // Re-borrow active_merge to write
-                    if let Some(merge) = &mut self.active_merge {
-                        merge.writer.write_partition(key, mutations)?;
-                        merge.rows_merged += row_count;
-                    }
-
-                    // Update stats
-                    report.rows_merged += row_count;
-                }
-                merge::MergeStep::Complete => {
-                    // Merge is complete - finalize and clean up
-                    // Use blocking call to handle async finalization
-                    self.finalize_merge_blocking(&mut report)?;
-                    break;
-                }
-            }
-        }
-
-        // Check if more work is pending
-        report.pending_compaction = self.active_merge.is_some();
-        report.time_spent = start.elapsed();
-
-        Ok(report)
-    }
-
-    /// Scan data directory for SSTable candidates (M5.2 helper)
-    /// Startup sweep (a): remove any `.compaction-tmp-*` directories left under
-    /// `data_dir` by a previous crash mid-rename.  Best-effort — individual
-    /// failures are logged but do not abort engine startup.
-    fn sweep_orphaned_compaction_tmp(data_dir: &Path) {
-        let read_dir = match std::fs::read_dir(data_dir) {
-            Ok(rd) => rd,
-            Err(e) => {
-                log::debug!(
-                    "sweep_orphaned_compaction_tmp: cannot read {:?}: {}",
-                    data_dir,
-                    e
-                );
-                return;
-            }
-        };
-
-        for entry in read_dir.flatten() {
-            let path = entry.path();
-            let name = entry.file_name();
-            let name_str = name.to_string_lossy();
-            if name_str.starts_with(".compaction-tmp-") && path.is_dir() {
-                log::warn!("removing orphaned compaction tmp directory: {:?}", path);
-                if let Err(e) = std::fs::remove_dir_all(&path) {
-                    log::warn!(
-                        "failed to remove orphaned compaction tmp directory {:?}: {}",
-                        path,
-                        e
-                    );
-                }
-            }
-        }
-    }
-
-    /// Startup sweep (b): remove any `nb-{gen}-big-Data.db` (and its siblings)
-    /// under `data_dir/keyspace/table/` that lack a matching `TOC.txt`.
-    /// Such files are left when a crash occurs after some component renames but
-    /// before `TOC.txt` is renamed (the publication barrier).  Best-effort.
-    fn sweep_orphaned_partial_sstables(data_dir: &Path, keyspace: &str, table: &str) {
-        let sstable_dir = data_dir.join(keyspace).join(table);
-
-        let read_dir = match std::fs::read_dir(&sstable_dir) {
-            Ok(rd) => rd,
-            Err(_) => {
-                // Directory doesn't exist yet — nothing to sweep.
-                return;
-            }
-        };
-
-        for entry in read_dir.flatten() {
-            let path = entry.path();
-            let name = entry.file_name();
-            let name_str = name.to_string_lossy();
-
-            // Look for Data.db files produced by the writer (nb-{gen}-big-Data.db)
-            if !name_str.starts_with("nb-")
-                || !name_str.ends_with("-big-Data.db")
-                || !path.is_file()
-            {
-                continue;
-            }
-
-            // Extract the base prefix (e.g. "nb-5-big") to find the TOC sibling
-            let base = match name_str.strip_suffix("-Data.db") {
-                Some(b) => b.to_owned(),
-                None => continue,
-            };
-
-            // Extract the generation number for the log message
-            let gen_str = base
-                .strip_prefix("nb-")
-                .and_then(|s| s.strip_suffix("-big"))
-                .unwrap_or(&base);
-
-            let toc_path = sstable_dir.join(format!("{}-TOC.txt", base));
-            if !toc_path.exists() {
-                log::warn!(
-                    "removing orphaned partial SSTable components for generation {}: missing TOC.txt",
-                    gen_str
-                );
-                if let Err(e) = Self::delete_sstable_files_static(&path) {
-                    log::warn!(
-                        "failed to remove orphaned partial SSTable for generation {}: {}",
-                        gen_str,
-                        e
-                    );
-                }
-            }
-        }
-    }
-
-    #[tracing::instrument(name = "compaction.scan_candidates", skip(self))]
-    fn scan_sstable_candidates(&self) -> Result<Vec<PathBuf>> {
-        let mut candidates = Vec::new();
-
-        if !self.config.data_dir.exists() {
-            return Ok(candidates);
-        }
-
-        Self::scan_data_files(
-            &self.config.data_dir,
-            &mut candidates,
-            crate::storage::sstable::MAX_SSTABLE_SCAN_DEPTH,
-        )?;
-        Ok(candidates)
-    }
-
-    /// Recursively scan for Data.db files
-    fn scan_data_files(dir: &Path, candidates: &mut Vec<PathBuf>, depth: usize) -> Result<()> {
-        for entry in std::fs::read_dir(dir)
-            .map_err(|e| Error::Storage(format!("Failed to read data directory: {}", e)))?
-        {
-            let entry = entry
-                .map_err(|e| Error::Storage(format!("Failed to read directory entry: {}", e)))?;
-
-            let path = entry.path();
-            let filename = path.file_name().unwrap_or_default().to_string_lossy();
-
-            // Only consider Data.db files
-            if filename.starts_with("nb-") && filename.ends_with("-big-Data.db") {
-                // Honor the TOC.txt publication barrier (Issue #591). A Data.db
-                // without a sibling TOC.txt is NOT a published SSTable: it is
-                // either a crash-interrupted partial rename or a deferred-delete
-                // orphan whose TOC was removed first while its data file stayed
-                // pinned by an open/mapped reader (Windows). Feeding such a file
-                // to the merger would re-compact an unpublished input and could
-                // produce garbled output, so it is skipped here just as the
-                // read path discovers SSTables by TOC.txt. The startup orphan
-                // sweep reclaims the leftover components.
-                let base = filename.trim_end_matches("-Data.db");
-                let toc_path = path.with_file_name(format!("{base}-TOC.txt"));
-                if toc_path.exists() {
-                    candidates.push(path);
-                } else {
-                    log::debug!(
-                        "scan_data_files: skipping unpublished SSTable (no TOC.txt): {:?}",
-                        path
-                    );
-                }
-            } else if depth > 0 && path.is_dir() {
-                Self::scan_data_files(&path, candidates, depth - 1)?;
-            }
-        }
-        Ok(())
-    }
-
-    /// Start a new merge operation (M5.2 helper, Issue #474)
-    ///
-    /// ## Atomicity design
-    ///
-    /// The SSTableWriter is pointed at a temporary directory (`tmp_dir`) that lives
-    /// alongside the final SSTable directory. After `writer.finish()` succeeds, each
-    /// output component is atomically renamed into the final directory. Input files are
-    /// deleted only after all renames complete. This guarantees:
-    ///
-    /// - A crash before renames: tmp files are incomplete; inputs intact.
-    /// - A crash mid-rename: at worst a partial output component exists in the final
-    ///   directory, but the old inputs are still there and the TOC.txt (publication
-    ///   barrier) has not been renamed yet, so the partial output is never discovered
-    ///   by readers scanning for `TOC.txt`.
-    /// - A crash after all renames but before input deletion: a harmless duplicate
-    ///   exists until next compaction.
-    ///
-    /// `purge_safe` (#921 finding 1): `true` only when `input_paths` spans every
-    /// SSTable for the table (a major/full compaction), allowing gc_grace
-    /// tombstone purging without risking resurrection of data shadowed in a
-    /// non-included overlapping SSTable. `false` keeps every tombstone.
-    ///
-    /// `max_purgeable_timestamp` (#935): for a PARTIAL compaction (`purge_safe ==
-    /// false`), the min write timestamp across the non-included overlapping
-    /// SSTables — see [`merge::compute_max_purgeable_timestamp`]. When `Some`, a
-    /// tombstone older than every outside SSTable is purged even in a partial
-    /// compaction; `None` keeps the conservative #921 behavior (no purging).
-    #[tracing::instrument(name = "compaction.start_merge", skip(self, input_paths, max_purgeable_timestamp), fields(inputs = input_paths.len()))]
-    fn start_merge(
-        &mut self,
-        input_paths: Vec<PathBuf>,
-        purge_safe: bool,
-        max_purgeable_timestamp: Option<i64>,
-    ) -> Result<()> {
-        log::info!(
-            "Starting compaction merge of {} SSTables",
-            input_paths.len()
-        );
-
-        // Measure total bytes read (sum of Data.db file sizes as an approximation)
-        let bytes_read: u64 = input_paths
-            .iter()
-            .map(|p| std::fs::metadata(p).map(|m| m.len()).unwrap_or(0))
-            .sum();
-
-        let output_generation = self.generation;
-
-        // Final SSTable directory: data_dir/keyspace/table/
-        let sstable_dir = self
-            .config
-            .data_dir
-            .join(&self.config.schema.keyspace)
-            .join(&self.config.schema.table);
-
-        // Temporary root: data_dir/.compaction-tmp-{gen}/
-        //
-        // Placing the tmp root inside `data_dir` (not inside `sstable_dir`) keeps
-        // the path simple and guarantees the rename is within the same filesystem.
-        // The SSTableWriter appends `keyspace/table/` internally, so component files
-        // land at: data_dir/.compaction-tmp-{gen}/keyspace/table/nb-{gen}-big-*.db
-        let tmp_dir = self
-            .config
-            .data_dir
-            .join(format!(".compaction-tmp-{}", output_generation));
-
-        // Create the tmp root (SSTableWriter::finish will create the subdirs)
-        std::fs::create_dir_all(&tmp_dir).map_err(|e| {
-            Error::Storage(format!(
-                "Failed to create compaction tmp directory {:?}: {}",
-                tmp_dir, e
-            ))
-        })?;
-
-        // #850: derive the effective compaction schema by reading static-row
-        // presence from the input SSTable headers. If a static column was dropped
-        // from the current schema but an input still carries it, re-add it so the
-        // merger decodes the static cells and the writer emits the static prelude
-        // (header-driven static presence). Byte-identical to `config.schema` when
-        // no such column exists.
-        let mut effective_schema =
-            merge::effective_compaction_schema(&self.config.schema, &input_paths);
-        // #929: normalize bare-name UDT columns to their UserType(...) marshal on
-        // the EFFECTIVE (decode) schema by copying each column's exact UserType
-        // marshal from the input headers. The compaction reader decides complex-
-        // vs-simple from this schema's `data_type` (is_complex_column), so a bare
-        // name would make it misdecode/drop pre-existing per-field UDT cells
-        // (roborev #1009/#1013/#1015/#1017/#1019/#1021). Derived ENTIRELY from the
-        // input headers, so it runs unconditionally (a registry is needed only at
-        // flush time); errors on a mixed encoding rather than corrupting.
-        // `write_schema` derives from this via `for_compaction_output`. The
-        // configured registry handles a UDT column added after the inputs were
-        // written (absent from every input header) so the output is not emitted
-        // as a bare/BytesType column (roborev #1023).
-        merge::apply_udt_marshals_from_inputs(
-            &mut effective_schema,
-            &input_paths,
-            self.config.udt_registry.as_ref(),
-        )?;
-
-        // Create K-way merger.
-        //
-        // gc_grace / gcBefore purging (issue #845): compute `gcBefore = now -
-        // gc_grace_seconds` from the table's `gc_grace_seconds` (read from the
-        // AUTHORITATIVE table schema's options, not the dropped-column-augmented
-        // `effective_schema`). When the table declares no `gc_grace_seconds`,
-        // `compute_gc_before` returns `None` and purging is a strict no-op,
-        // preserving the pre-#845 behavior. `now_secs` is the wall-clock used for
-        // both the purge boundary and TTL evaluation.
-        let now_secs = std::time::SystemTime::now()
-            .duration_since(std::time::UNIX_EPOCH)
-            .map(|d| d.as_secs() as i64)
-            .unwrap_or(0);
-        let gc_before_secs = merge::compute_gc_before(&self.config.schema, now_secs);
-        // Overlap-safety gate (#921 finding 1): only a major/full compaction
-        // (one that spans every candidate SSTable for the table) may purge
-        // tombstones. A partial compaction sets `purge_safe = false`, which
-        // makes the gc_grace purge stage a strict no-op and cannot resurrect
-        // data shadowed in a non-included overlapping SSTable.
-
-        // #921 finding 2: mirror `compact_sstables`' dropped-column survivor
-        // pre-pass. Decode with `effective_schema` (retains dropped columns so
-        // their cells parse and can be purged), but WRITE with a post-drop
-        // schema: a background compaction that purges the LAST cell/tombstone of
-        // a dropped column must NOT emit that column in the output
-        // SerializationHeader (stale header columns misalign a post-drop reader).
-        // The pre-pass uses the SAME gc cutoff AND `purge_safe` as the write
-        // merger below so the two make IDENTICAL purge decisions — a tombstone
-        // purged in the write pass is also purged here and never counted as a
-        // surviving cell. Reuses the exact helper `compact_sstables` uses.
-        let retained_dropped = if effective_schema.dropped_columns.is_empty() {
-            std::collections::HashSet::new()
-        } else {
-            merge::compute_surviving_dropped_columns(
-                input_paths.clone(),
-                &effective_schema,
-                gc_before_secs,
-                Some(now_secs),
-                purge_safe,
-                max_purgeable_timestamp,
-            )?
-        };
-        // `write_schema` inherits the #929 UDT normalization from the already-
-        // normalized `effective_schema` (for_compaction_output clones columns).
-        let write_schema = effective_schema.for_compaction_output(&retained_dropped);
-
-        let merger = KWayMerger::new_with_gc(
-            input_paths.clone(),
-            &effective_schema,
-            gc_before_secs,
-            Some(now_secs),
-        )?
-        .with_purge_safe(purge_safe)
-        // #935: overlap-aware purging for a partial compaction. `purge_safe`
-        // overrides this in the merger (full compaction → +inf bound), so passing
-        // the partial bound unconditionally is safe.
-        .with_max_purgeable_timestamp(max_purgeable_timestamp);
-
-        // Point the SSTableWriter at the tmp root; it will write to
-        // tmp_dir/keyspace/table/nb-{gen}-big-*.db. Use the post-drop
-        // `write_schema` (NOT `effective_schema`) so fully-purged dropped
-        // columns are stripped from the output header (#921 finding 2).
-        let mut writer = crate::storage::sstable::writer::SSTableWriter::new(
-            tmp_dir.clone(),
-            output_generation,
-            &write_schema,
-        )?;
-
-        // Two-pass compaction (issue #729): compute output FINAL encoding baselines
-        // by reading the minimum values from each input SSTable's Statistics.db.
-        // The output baseline must be ≤ all per-partition values from all inputs.
-        let (baseline_min_ts, baseline_min_ldt, baseline_min_ttl) =
-            merge::compute_baseline_min(&input_paths);
-        writer.pre_seed_encoding_baselines(baseline_min_ts, baseline_min_ldt, baseline_min_ttl);
-
-        // Increment generation for next operation
-        self.generation += 1;
-
-        self.active_merge = Some(ActiveMerge {
-            merger,
-            writer,
-            input_paths,
-            tmp_dir,
-            sstable_dir,
-            rows_merged: 0,
-            bytes_read,
-            started_at: Instant::now(),
-            effective_schema,
-        });
-
-        Ok(())
-    }
-
-    /// Finalize the active merge - blocking version (M5.2 helper).
-    ///
-    /// Bridges to the async finalizer via [`merge::block_on_async`], which is
-    /// safe to call from within an active Tokio runtime (e.g. the CLI's
-    /// `#[tokio::main]` worker threads). A nested `Handle::block_on` here would
-    /// otherwise panic with "Cannot start a runtime from within a runtime"
-    /// (Issue #587).
-    fn finalize_merge_blocking(&mut self, report: &mut MaintenanceReport) -> Result<()> {
-        merge::block_on_async(self.finalize_merge_async(report))
-    }
-
-    /// Finalize the active merge - async version (M5.2 helper, Issue #474)
-    ///
-    /// ## Atomicity protocol
-    ///
-    /// 1. `writer.finish()` flushes all component files to the tmp directory.
-    /// 2. Each component file is renamed from `tmp_dir/` to `sstable_dir/`.
-    ///    The TOC.txt rename is performed **last** (it is the publication barrier:
-    ///    readers discover SSTables by scanning for `TOC.txt`).
-    /// 3. Only after all renames succeed are the input SSTable files deleted.
-    /// 4. The now-empty tmp directory is removed.
-    ///
-    /// If any step 2 rename fails, the partially-renamed output components are
-    /// cleaned up and an error is returned. The input SSTables remain intact.
-    ///
-    /// Single-boundary error recording (issue #1037): this is an *unrecorded*
-    /// inner helper. Any error returned here escapes through
-    /// `maintenance_step_inner` to the public `maintenance_step`, which wraps the
-    /// whole step in `record_result("compaction", ..)` and counts it exactly
-    /// once. Recording here too would double-count finalize failures.
-    #[tracing::instrument(name = "compaction.finalize", skip(self, report))]
-    async fn finalize_merge_async(&mut self, report: &mut MaintenanceReport) -> Result<()> {
-        let merge = match self.active_merge.take() {
-            Some(m) => m,
-            None => return Ok(()),
-        };
-
-        let input_count = merge.input_paths.len() as u64;
-        let merge_rows = merge.rows_merged;
-        let elapsed = merge.started_at.elapsed();
-        log::info!(
-            "Finalizing compaction merge: {} rows, {:?} elapsed",
-            merge.rows_merged,
-            elapsed
-        );
-
-        // Step 1: Finish writing all components to the tmp directory.
-        // If this fails the tmp directory may contain partial files, but inputs are safe.
-        let tmp_info = match merge.writer.finish().await {
-            Ok(info) => info,
-            Err(e) => {
-                // Clean up tmp directory on failure (best effort)
-                let _ = std::fs::remove_dir_all(&merge.tmp_dir);
-                return Err(Error::Storage(format!(
-                    "Compaction merge write failed (inputs intact): {}",
-                    e
-                )));
-            }
-        };
-
-        log::info!(
-            "Compaction tmp output: {} bytes, {} partitions",
-            tmp_info.data_size,
-            tmp_info.partition_count
-        );
-
-        // Step 2: Atomically rename each component from the tmp sub-directory to
-        // the final SSTable directory. Because both directories are under the same
-        // `data_dir`, the rename is within the same filesystem (POSIX atomic).
-        // We rename TOC.txt last because it is the publication barrier.
-        let sstable_dir = &merge.sstable_dir;
-
-        // Ensure the final SSTable directory exists (it normally does, but handle
-        // the edge case where it was created by start_merge only in the tmp path).
-        std::fs::create_dir_all(sstable_dir).map_err(|e| {
-            Error::Storage(format!(
-                "Failed to create SSTable directory {:?}: {}",
-                sstable_dir, e
-            ))
-        })?;
-
-        // Helper: map a tmp component path to its final destination.
-        // tmp_info paths look like: data_dir/.compaction-tmp-N/keyspace/table/nb-N-big-X.db
-        // Final destination:        data_dir/keyspace/table/nb-N-big-X.db
-        let make_rename = |src: &PathBuf| -> Result<(PathBuf, PathBuf)> {
-            let filename = src
-                .file_name()
-                .ok_or_else(|| Error::Storage("Component path has no filename".to_string()))?;
-            let dst = sstable_dir.join(filename);
-            Ok((src.clone(), dst))
-        };
-
-        // Build list of (src, dst) renames. TOC.txt goes last (publication barrier).
-        let mut renames: Vec<(PathBuf, PathBuf)> = Vec::new();
-
-        // Non-TOC components first. Index.db/Summary.db are BIG-only (BTI writers
-        // report them as None — issue #908); only rename them when present.
-        // Filter.db is optional too: a table with a disabled bloom filter
-        // (bloom_filter_fp_chance = 1.0, AlwaysPresentFilter) emits NO Filter.db,
-        // so there is nothing to rename. Including a non-existent path here would
-        // fail the atomic publish (Issue #852).
-        let mut non_toc: Vec<&PathBuf> = vec![&tmp_info.data_path];
-        if let Some(ref p) = tmp_info.index_path {
-            non_toc.push(p);
-        }
-        if let Some(ref p) = tmp_info.filter_path {
-            non_toc.push(p);
-        }
-        if let Some(ref p) = tmp_info.summary_path {
-            non_toc.push(p);
-        }
-        non_toc.push(&tmp_info.stats_path);
-        non_toc.push(&tmp_info.digest_path);
-        for src in non_toc {
-            renames.push(make_rename(src)?);
-        }
-        if let Some(ref p) = tmp_info.partitions_path {
-            renames.push(make_rename(p)?);
-        }
-        if let Some(ref ci_path) = tmp_info.compression_info_path {
-            renames.push(make_rename(ci_path)?);
-        }
-        // TOC.txt is last (publication barrier)
-        renames.push(make_rename(&tmp_info.toc_path)?);
-
-        // Perform the renames. On failure, remove any already-renamed files so
-        // we don't leave a half-published SSTable, then return the error.
-        // Time the rename/publication-barrier phase (issue #1037).
-        let finalize_start = Instant::now();
-        let mut renamed: Vec<PathBuf> = Vec::with_capacity(renames.len());
-        let mut rename_error: Option<Error> = None;
-
-        for (src, dst) in &renames {
-            match std::fs::rename(src, dst) {
-                Ok(()) => {
-                    log::debug!(
-                        "Renamed {:?} → {:?}",
-                        src.file_name().unwrap_or_default(),
-                        dst.file_name().unwrap_or_default()
-                    );
-                    renamed.push(dst.clone());
-                }
-                Err(e) => {
-                    rename_error = Some(Error::Storage(format!(
-                        "Atomic rename of {:?} to {:?} failed (rolling back, inputs intact): {}",
-                        src, dst, e
-                    )));
-                    break;
-                }
-            }
-        }
-
-        if let Some(err) = rename_error {
-            // Roll back already-renamed files (best effort)
-            for dst in &renamed {
-                let _ = std::fs::remove_file(dst);
-            }
-            // Clean up tmp directory
-            let _ = std::fs::remove_dir_all(&merge.tmp_dir);
-            return Err(err);
-        }
-
-        // Finalize/rename latency in seconds (issue #1037): the atomic
-        // publication barrier just completed successfully.
-        crate::observability::record_histogram(
-            crate::observability::catalog::COMPACTION_FINALIZE_DURATION,
-            finalize_start.elapsed().as_secs_f64(),
-            &[],
-        );
-
-        // Step 3: All renames succeeded. The new SSTable is now visible.
-        // Delete input SSTable files. If deletion fails we log a warning but do
-        // NOT return an error — the merge output is correct.
-        //
-        // Issue #591 (write-while-mapped / Windows policy): the inputs were read
-        // through buffered I/O (never memory-mapped) and the merge has fully
-        // drained every source through its streaming producer by this point
-        // (issue #827), so the merger holds no mapping over them — deleting them
-        // cannot fault with SIGBUS. `delete_sstable_files` removes each
-        // input's TOC.txt first (unpublishing it) and is best-effort on the data
-        // components, so a component still pinned by a concurrent mapped reader on
-        // Windows becomes an invisible orphan reclaimed on the next startup rather
-        // than a hard failure or a source of duplicate rows.
-        for input_path in &merge.input_paths {
-            if let Err(e) = self.delete_sstable_files(input_path) {
-                log::warn!(
-                    "Failed to delete compaction input {:?}: {} \
-                     (merge output is valid; inputs will be re-evaluated next cycle)",
-                    input_path,
-                    e
-                );
-            }
-        }
-
-        // Step 4: Remove the now-empty tmp directory (best effort).
-        if let Err(e) = std::fs::remove_dir_all(&merge.tmp_dir) {
-            log::debug!(
-                "Failed to remove compaction tmp directory {:?}: {}",
-                merge.tmp_dir,
-                e
-            );
-        }
-
-        // The final Data.db path is in sstable_dir (renamed from tmp)
-        let final_data_path = sstable_dir.join(
-            tmp_info
-                .data_path
-                .file_name()
-                .ok_or_else(|| Error::Storage("Data.db path has no filename".to_string()))?,
-        );
-
-        // Compute total bytes written across ALL output SSTable components (not just Data.db).
-        // We stat the final paths (post-rename) so we measure what was actually persisted.
-        let stat_final = |p: &PathBuf| -> u64 {
-            let filename = p.file_name().unwrap_or_default();
-            std::fs::metadata(sstable_dir.join(filename))
-                .map(|m| m.len())
-                .unwrap_or(0)
-        };
-        // Index.db/Summary.db are BIG-only (BTI reports None — issue #908) and
-        // Filter.db is optional (disabled bloom filter omits it, Issue #852), so
-        // only stat the components that were actually written.
-        let mut byte_paths: Vec<&PathBuf> = vec![&tmp_info.data_path];
-        if let Some(ref p) = tmp_info.index_path {
-            byte_paths.push(p);
-        }
-        if let Some(ref p) = tmp_info.filter_path {
-            byte_paths.push(p);
-        }
-        if let Some(ref p) = tmp_info.summary_path {
-            byte_paths.push(p);
-        }
-        byte_paths.push(&tmp_info.stats_path);
-        byte_paths.push(&tmp_info.digest_path);
-        if let Some(ref p) = tmp_info.partitions_path {
-            byte_paths.push(p);
-        }
-        let total_bytes_written: u64 = byte_paths.iter().map(|p| stat_final(p)).sum::<u64>()
-            + tmp_info
-                .compression_info_path
-                .as_ref()
-                .map(stat_final)
-                .unwrap_or(0);
-
-        // Update per-step report
-        report.completed_merges.push(final_data_path);
-        report.bytes_written += total_bytes_written;
-
-        // Update cumulative lifetime stats
-        self.cumulative_stats.compactions_completed += 1;
-        self.cumulative_stats.sstables_merged_in += merge.input_paths.len() as u64;
-        self.cumulative_stats.sstables_produced += 1;
-        self.cumulative_stats.bytes_read += merge.bytes_read;
-        self.cumulative_stats.bytes_written += total_bytes_written;
-        self.cumulative_stats.rows_merged += merge.rows_merged;
-        self.cumulative_stats.total_time += elapsed;
-
-        // Compaction completion metrics (issue #1037): full-compaction duration
-        // (pairs with COMPACTION_ROWS_MERGED — emitted incrementally per
-        // maintenance step — for rows/sec throughput), bytes written across all
-        // output components, and SSTables in/out. Rows are NOT re-counted here to
-        // avoid double-counting the per-step COMPACTION_ROWS_MERGED increments.
-        let _ = merge_rows;
-        {
-            use crate::observability::{self as obs, catalog};
-            obs::record_histogram(catalog::COMPACTION_DURATION, elapsed.as_secs_f64(), &[]);
-            obs::add_counter(catalog::COMPACTION_BYTES_WRITTEN, total_bytes_written, &[]);
-            obs::add_counter(catalog::COMPACTION_SSTABLES_IN, input_count, &[]);
-            obs::add_counter(catalog::COMPACTION_SSTABLES_OUT, 1, &[]);
-        }
-
-        log::info!(
-            "Compaction complete: merged {} inputs → 1 output ({} bytes total across all components, {} rows, {:?})",
-            merge.input_paths.len(),
-            total_bytes_written,
-            merge.rows_merged,
-            elapsed
-        );
-
-        Ok(())
-    }
-
-    /// Delete all component files for an SSTable (M5.2 helper)
-    fn delete_sstable_files(&self, data_path: &Path) -> Result<()> {
-        Self::delete_sstable_files_static(data_path)
-    }
-
-    /// Static helper that deletes all component files for an SSTable given the
-    /// Data.db path.  Called from both `delete_sstable_files` and the startup
-    /// orphan sweep, which runs before `self` is fully constructed.
-    ///
-    /// ## Deferred-delete / Windows policy (Issue #591)
-    ///
-    /// `TOC.txt` is removed **first**. TOC.txt is the publication barrier — both
-    /// the read path (`SSTableManager`) and the compaction candidate scan
-    /// (`scan_data_files`, since #591) treat a Data.db without a sibling TOC.txt
-    /// as unpublished. Removing TOC.txt first therefore *unpublishes* the SSTable
-    /// atomically, before any data component is touched, so it can never be
-    /// observed (no duplicate rows, never re-fed to the merger) even if the
-    /// remaining components cannot be removed yet.
-    ///
-    /// The remaining components are then deleted **best-effort**: a failure on
-    /// any one of them (most plausibly a Windows sharing violation when a
-    /// concurrent reader still has the file open or memory-mapped) is logged but
-    /// does NOT abort the rest or fail the operation. Such a leftover is a
-    /// harmless orphan — invisible because its TOC.txt is gone — and is reclaimed
-    /// by [`Self::sweep_orphaned_partial_sstables`] on the next engine startup,
-    /// by which time the reader's handle has been released. This is the
-    /// "deferred delete" half of the policy; Unix removes the inode immediately
-    /// while any mapping keeps the bytes alive until it is dropped.
-    fn delete_sstable_files_static(data_path: &Path) -> Result<()> {
-        // Extract base path: nb-{gen}-big
-        let filename = data_path
-            .file_name()
-            .and_then(|s| s.to_str())
-            .ok_or_else(|| Error::Storage("Invalid SSTable path".to_string()))?;
-
-        let base = filename
-            .strip_suffix("-Data.db")
-            .ok_or_else(|| Error::Storage("Invalid Data.db filename".to_string()))?;
-
-        let parent_dir = data_path.parent().ok_or_else(|| {
-            Error::Storage(format!(
-                "Data.db path has no parent directory: {:?}",
-                data_path
-            ))
-        })?;
-
-        // TOC.txt FIRST — the publication barrier (Issue #591). Once it is gone
-        // the SSTable is unpublished regardless of whether the data components
-        // can be removed. Remaining components follow, best-effort.
-        let components = [
-            "TOC.txt",
-            "Data.db",
-            "Index.db",
-            "Summary.db",
-            "Statistics.db",
-            "CompressionInfo.db",
-            "Filter.db",
-            "Digest.crc32",
-        ];
-
-        let mut failures: Vec<String> = Vec::new();
-        for component in &components {
-            let component_path = parent_dir.join(format!("{}-{}", base, component));
-            if component_path.exists() {
-                match std::fs::remove_file(&component_path) {
-                    Ok(()) => log::debug!("Deleted compaction input: {:?}", component_path),
-                    Err(e) => {
-                        // Best-effort: do not abort. A leftover data component
-                        // whose TOC.txt is already gone is an invisible orphan
-                        // reclaimed by the startup sweep (Issue #591).
-                        log::warn!(
-                            "Deferred delete of {:?}: {} (component left as orphan; \
-                             unpublished via TOC.txt removal, reclaimed on next startup)",
-                            component_path,
-                            e
-                        );
-                        failures.push(format!("{:?}: {}", component_path, e));
-                    }
-                }
-            }
-        }
-
-        if failures.is_empty() {
-            Ok(())
-        } else {
-            // Surface a non-fatal error so callers can log it. The SSTable is
-            // already unpublished (TOC.txt removed first), so callers treat this
-            // as a deferred reclamation, not a correctness failure.
-            Err(Error::Storage(format!(
-                "Deferred delete left {} orphaned component(s) (unpublished, reclaimed on \
-                 next startup): {}",
-                failures.len(),
-                failures.join("; ")
-            )))
-        }
-    }
 }
 
 /// Safety-net `Drop` implementation for `WriteEngine`.
@@ -2182,78 +1027,8 @@ impl Drop for WriteEngine {
 #[cfg(all(test, feature = "write-support"))]
 mod tests {
     use super::*;
-    use crate::schema::{Column, KeyColumn};
-    use crate::storage::write_engine::mutation::{CellOperation, PartitionKey, TableId};
-    use crate::types::Value;
-    use std::collections::HashMap;
+    use crate::storage::write_engine::test_support::{create_test_mutation, create_test_schema};
     use tempfile::TempDir;
-
-    fn create_test_schema() -> TableSchema {
-        TableSchema {
-            keyspace: "test_ks".to_string(),
-            table: "test_table".to_string(),
-            partition_keys: vec![KeyColumn {
-                name: "id".to_string(),
-                data_type: "int".to_string(),
-                position: 0,
-            }],
-            clustering_keys: vec![],
-            columns: vec![
-                Column {
-                    name: "id".to_string(),
-                    data_type: "int".to_string(),
-                    nullable: false,
-                    default: None,
-                    is_static: false,
-                },
-                Column {
-                    name: "name".to_string(),
-                    data_type: "text".to_string(),
-                    nullable: true,
-                    default: None,
-                    is_static: false,
-                },
-            ],
-            comments: HashMap::new(),
-            dropped_columns: HashMap::new(),
-        }
-    }
-
-    fn create_test_mutation(id: i32, name: &str, timestamp: i64) -> Mutation {
-        let table_id = TableId::new("test_ks", "test_table");
-        let pk = PartitionKey::single("id", Value::Integer(id));
-        let ops = vec![CellOperation::Write {
-            column: "name".to_string(),
-            value: Value::Text(name.to_string()),
-        }];
-
-        Mutation::new(table_id, pk, None, ops, timestamp, None)
-    }
-
-    #[test]
-    fn test_set_merge_policy() {
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Should succeed now (was previously returning error)
-        let policy = Box::new(crate::storage::write_engine::STCSPolicy::default());
-        engine.set_merge_policy(policy).unwrap();
-
-        // With policy set but no SSTables, should return quickly with no work
-        let report = engine
-            .maintenance_step(std::time::Duration::from_millis(100))
-            .unwrap();
-        assert!(!report.pending_compaction);
-        assert_eq!(report.rows_merged, 0);
-    }
 
     #[test]
     fn test_write_engine_config() {
@@ -2969,955 +1744,6 @@ mod tests {
         assert!(generation > u32::MAX as u64);
     }
 
-    // M5.2 maintenance_step() tests (Issue #384)
-
-    #[test]
-    fn test_maintenance_step_no_policy() {
-        // Without a merge policy, maintenance_step should do nothing
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Call maintenance_step without setting a policy
-        let report = engine.maintenance_step(Duration::from_millis(100)).unwrap();
-
-        // Should return immediately with no work done
-        assert_eq!(report.rows_merged, 0);
-        assert_eq!(report.bytes_written, 0);
-        assert_eq!(report.completed_merges.len(), 0);
-        assert!(!report.pending_compaction);
-        assert!(report.time_spent < Duration::from_millis(50));
-    }
-
-    #[test]
-    fn test_maintenance_step_with_closed_engine() {
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Close the engine
-        tokio::runtime::Runtime::new()
-            .unwrap()
-            .block_on(engine.close())
-            .unwrap();
-
-        // maintenance_step should fail on closed engine
-        let result = engine.maintenance_step(Duration::from_millis(100));
-        assert!(result.is_err());
-        match result {
-            Err(Error::InvalidInput(msg)) => {
-                assert!(msg.contains("closed"));
-            }
-            _ => panic!("Expected InvalidInput error"),
-        }
-    }
-
-    #[test]
-    fn test_maintenance_report_creation() {
-        let report = MaintenanceReport {
-            time_spent: Duration::from_millis(250),
-            completed_merges: vec![PathBuf::from("data/nb-5-big-Data.db")],
-            rows_merged: 1000,
-            bytes_written: 1024 * 1024,
-            pending_compaction: true,
-        };
-
-        assert_eq!(report.time_spent.as_millis(), 250);
-        assert_eq!(report.completed_merges.len(), 1);
-        assert_eq!(report.rows_merged, 1000);
-        assert_eq!(report.bytes_written, 1024 * 1024);
-        assert!(report.pending_compaction);
-    }
-
-    #[test]
-    fn test_scan_sstable_candidates_empty_dir() {
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let engine = WriteEngine::new(config).unwrap();
-
-        let candidates = engine.scan_sstable_candidates().unwrap();
-        assert_eq!(candidates.len(), 0);
-    }
-
-    #[test]
-    fn test_scan_sstable_candidates_with_sstables() {
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let engine = WriteEngine::new(config).unwrap();
-
-        // Create dummy SSTable files. Each Data.db needs a sibling TOC.txt to
-        // count as a *published* SSTable (the publication barrier, Issue #591) —
-        // a Data.db without TOC.txt is an unpublished partial/orphan and must be
-        // skipped by the candidate scan.
-        let data_dir = temp_dir.path().join("data");
-        std::fs::create_dir_all(&data_dir).unwrap();
-        std::fs::write(data_dir.join("nb-1-big-Data.db"), b"").unwrap();
-        std::fs::write(data_dir.join("nb-1-big-TOC.txt"), b"").unwrap();
-        std::fs::write(data_dir.join("nb-2-big-Data.db"), b"").unwrap();
-        std::fs::write(data_dir.join("nb-2-big-TOC.txt"), b"").unwrap();
-        std::fs::write(data_dir.join("nb-3-big-Index.db"), b"").unwrap(); // Not a Data.db
-        std::fs::write(data_dir.join("other-file.txt"), b"").unwrap(); // Not an SSTable
-                                                                       // An unpublished Data.db (no TOC.txt) must NOT be picked up (Issue #591).
-        std::fs::write(data_dir.join("nb-4-big-Data.db"), b"").unwrap();
-
-        let candidates = engine.scan_sstable_candidates().unwrap();
-
-        // Should only find the two PUBLISHED Data.db files (TOC.txt present);
-        // nb-4 is excluded because it has no TOC.txt.
-        assert_eq!(candidates.len(), 2);
-        assert!(candidates
-            .iter()
-            .all(|p| p.to_string_lossy().contains("Data.db")));
-        assert!(
-            !candidates
-                .iter()
-                .any(|p| p.to_string_lossy().contains("nb-4-big")),
-            "unpublished Data.db (no TOC.txt) must be excluded (Issue #591)"
-        );
-    }
-
-    #[test]
-    fn test_delete_sstable_files() {
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let engine = WriteEngine::new(config).unwrap();
-
-        // Create dummy SSTable component files
-        let data_dir = temp_dir.path().join("data");
-        std::fs::create_dir_all(&data_dir).unwrap();
-
-        let components = [
-            "nb-5-big-Data.db",
-            "nb-5-big-Index.db",
-            "nb-5-big-Summary.db",
-            "nb-5-big-Statistics.db",
-        ];
-
-        for component in &components {
-            std::fs::write(data_dir.join(component), b"dummy").unwrap();
-        }
-
-        // Verify files exist
-        for component in &components {
-            assert!(data_dir.join(component).exists());
-        }
-
-        // Delete SSTable files
-        let data_path = data_dir.join("nb-5-big-Data.db");
-        engine.delete_sstable_files(&data_path).unwrap();
-
-        // Verify files are deleted
-        for component in &components {
-            assert!(!data_dir.join(component).exists());
-        }
-    }
-
-    /// Issue #591: deletion removes TOC.txt FIRST so the SSTable is unpublished
-    /// before any data component is touched. This guarantees the read path and
-    /// the compaction candidate scan stop seeing it immediately, even if a data
-    /// component cannot be removed yet (e.g. pinned by a mapped reader on
-    /// Windows).
-    #[test]
-    fn test_delete_removes_toc_first_unpublishing_atomically() {
-        let temp_dir = TempDir::new().unwrap();
-        let data_dir = temp_dir.path().join("data");
-        std::fs::create_dir_all(&data_dir).unwrap();
-
-        // A full published SSTable component set including TOC.txt.
-        for comp in &[
-            "nb-7-big-Data.db",
-            "nb-7-big-Index.db",
-            "nb-7-big-Statistics.db",
-            "nb-7-big-TOC.txt",
-        ] {
-            std::fs::write(data_dir.join(comp), b"x").unwrap();
-        }
-
-        let data_path = data_dir.join("nb-7-big-Data.db");
-        WriteEngine::delete_sstable_files_static(&data_path).unwrap();
-
-        // Everything gone on the happy path.
-        assert!(!data_dir.join("nb-7-big-TOC.txt").exists());
-        assert!(!data_path.exists());
-
-        // And critically: scan_data_files (the compaction candidate discovery)
-        // never surfaces a Data.db without a TOC.txt, so a deferred-delete orphan
-        // is not re-fed to the merger. Recreate a TOC-less leftover to prove it.
-        std::fs::write(data_dir.join("nb-8-big-Data.db"), b"x").unwrap();
-        let mut candidates = Vec::new();
-        WriteEngine::scan_data_files(&data_dir, &mut candidates, 1).unwrap();
-        assert!(
-            candidates.is_empty(),
-            "a Data.db without a sibling TOC.txt must NOT be a compaction candidate \
-             (publication barrier, Issue #591); got {:?}",
-            candidates
-        );
-
-        // Add the matching TOC.txt and it becomes a valid candidate again.
-        std::fs::write(data_dir.join("nb-8-big-TOC.txt"), b"x").unwrap();
-        let mut candidates = Vec::new();
-        WriteEngine::scan_data_files(&data_dir, &mut candidates, 1).unwrap();
-        assert_eq!(
-            candidates.len(),
-            1,
-            "a published Data.db (TOC.txt present) must be discovered"
-        );
-    }
-
-    // Mock merge policy that selects specific files for testing
-    #[derive(Debug)]
-    #[allow(dead_code)] // Used in multiple test functions below
-    struct TestMergePolicy {
-        files_to_select: Vec<PathBuf>,
-    }
-
-    impl MergePolicy for TestMergePolicy {
-        fn select_merge(&self, _candidates: &[PathBuf]) -> Result<Vec<PathBuf>> {
-            Ok(self.files_to_select.clone())
-        }
-    }
-
-    #[test]
-    fn test_maintenance_step_with_policy_no_work() {
-        // Policy that returns empty selection (no work to do)
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Set a policy that selects nothing
-        let policy = TestMergePolicy {
-            files_to_select: vec![],
-        };
-        engine.set_merge_policy(Box::new(policy)).unwrap();
-
-        // Call maintenance_step - policy selects no work
-        let report = engine.maintenance_step(Duration::from_millis(100)).unwrap();
-
-        // Should return with no work done
-        assert_eq!(report.rows_merged, 0);
-        assert_eq!(report.bytes_written, 0);
-        assert_eq!(report.completed_merges.len(), 0);
-        assert!(!report.pending_compaction);
-    }
-
-    #[test]
-    fn test_maintenance_step_budget_honored() {
-        // Test that budget is approximately honored
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Set a policy that selects nothing
-        let policy = TestMergePolicy {
-            files_to_select: vec![],
-        };
-        engine.set_merge_policy(Box::new(policy)).unwrap();
-
-        // Call with small budget - policy selects no work, should return quickly
-        let budget = Duration::from_millis(10);
-        let report = engine.maintenance_step(budget).unwrap();
-
-        // Should return quickly when there's no compaction work
-        assert!(
-            report.time_spent < budget.mul_f32(1.5),
-            "Time spent {:?} exceeded budget {:?} by >50%",
-            report.time_spent,
-            budget
-        );
-    }
-
-    // ============================================================================
-    // Issue #474: Wire k-way merger + STCS into maintenance_step
-    // ============================================================================
-
-    /// Helper: flush `n` distinct mutations through the engine synchronously.
-    ///
-    /// Uses a dedicated single-threaded runtime so it can be called from both
-    /// sync test functions and (via `spawn_blocking`) from async contexts.
-    fn flush_n_sstables_sync(engine: &mut WriteEngine, n: usize) -> Vec<PathBuf> {
-        let rt = tokio::runtime::Builder::new_current_thread()
-            .enable_all()
-            .build()
-            .unwrap();
-        let mut paths = Vec::new();
-        for batch in 0..n {
-            for row in 0..5 {
-                let mutation = create_test_mutation(
-                    (batch * 100 + row) as i32,
-                    &format!("User-{}-{}", batch, row),
-                    1_000_000 + (batch * 100 + row) as i64,
-                );
-                engine.write(mutation).unwrap();
-            }
-            let info = rt.block_on(engine.flush()).unwrap().unwrap();
-            paths.push(info.data_path);
-        }
-        paths
-    }
-
-    #[test]
-    fn test_maintenance_stats_initial_zero() {
-        // Before any maintenance work, all stats should be zero
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let engine = WriteEngine::new(config).unwrap();
-
-        let stats = engine.maintenance_stats();
-        assert_eq!(stats.compactions_completed, 0);
-        assert_eq!(stats.sstables_merged_in, 0);
-        assert_eq!(stats.sstables_produced, 0);
-        assert_eq!(stats.bytes_read, 0);
-        assert_eq!(stats.bytes_written, 0);
-        assert_eq!(stats.rows_merged, 0);
-        assert_eq!(stats.total_time, Duration::ZERO);
-    }
-
-    #[test]
-    fn test_stcs_selects_expected_group_by_size() {
-        // Verify that STCSPolicy groups four same-sized SSTables into one candidate set.
-        // We do this without actually running a merge (just test the policy selection).
-        let policy = crate::storage::write_engine::STCSPolicy::default();
-
-        // Create 4 temp files of equal size to satisfy min_threshold=4
-        let temp_dir = TempDir::new().unwrap();
-        let mut paths = Vec::new();
-        for i in 1..=4 {
-            let path = temp_dir.path().join(format!("nb-{}-big-Data.db", i));
-            // 60 MB each (above min_sstable_size threshold)
-            let size_bytes = 60 * 1024 * 1024u64;
-            let file = std::fs::File::create(&path).unwrap();
-            file.set_len(size_bytes).unwrap();
-            paths.push(path);
-        }
-
-        // Policy should select all 4 as a candidate group
-        let selected = policy.select_merge(&paths).unwrap();
-        assert_eq!(
-            selected.len(),
-            4,
-            "STCS should select all 4 same-sized SSTables as one compaction group"
-        );
-
-        // All selected paths should be from our input set
-        for sel in &selected {
-            assert!(
-                paths.contains(sel),
-                "Selected path {:?} not in input set",
-                sel
-            );
-        }
-    }
-
-    #[test]
-    fn test_stcs_does_not_select_below_threshold() {
-        // With only 3 SSTables, STCS (min_threshold=4) should select nothing.
-        let policy = crate::storage::write_engine::STCSPolicy::default();
-
-        let temp_dir = TempDir::new().unwrap();
-        let mut paths = Vec::new();
-        for i in 1..=3 {
-            let path = temp_dir.path().join(format!("nb-{}-big-Data.db", i));
-            let file = std::fs::File::create(&path).unwrap();
-            file.set_len(60 * 1024 * 1024).unwrap();
-            paths.push(path);
-        }
-
-        let selected = policy.select_merge(&paths).unwrap();
-        assert!(
-            selected.is_empty(),
-            "STCS should NOT select when fewer than min_threshold SSTables exist"
-        );
-    }
-
-    #[test]
-    fn test_maintenance_step_compacts_sstables_atomically() {
-        // Create an engine, flush 4 SSTables, then run maintenance_step with STCS.
-        // After the step: input files must be gone, output file must exist,
-        // and maintenance_stats() must reflect the completed compaction.
-        //
-        // Uses a sync wrapper so maintenance_step's internal block_on works without
-        // nesting inside a pre-existing async runtime.
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        // Use a LOW min_sstable_size so small test files pass bucket grouping
-        let policy = crate::storage::write_engine::STCSPolicy::new(
-            4,   // min_threshold
-            32,  // max_threshold
-            0.5, // bucket_low
-            1.5, // bucket_high
-            0,   // min_sstable_size = 0 so tiny files group together
-        )
-        .unwrap();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Flush 4 distinct SSTables (sync helper creates its own single-threaded runtime)
-        let input_paths = flush_n_sstables_sync(&mut engine, 4);
-        assert_eq!(input_paths.len(), 4, "Expected 4 flushed SSTables");
-
-        // Verify all input Data.db files exist before compaction
-        for p in &input_paths {
-            assert!(
-                p.exists(),
-                "Input file {:?} should exist before compaction",
-                p
-            );
-        }
-
-        // Attach the policy and run maintenance
-        engine.set_merge_policy(Box::new(policy)).unwrap();
-        let report = engine.maintenance_step(Duration::from_secs(60)).unwrap();
-
-        // The report must indicate a completed merge
-        assert_eq!(
-            report.completed_merges.len(),
-            1,
-            "Expected exactly 1 completed merge, got: {:?}",
-            report.completed_merges
-        );
-        // bytes_written is u64 and always non-negative, so no assertion needed here.
-
-        // The merged output file must exist in the final SSTable directory
-        let merged_path = &report.completed_merges[0];
-        assert!(
-            merged_path.exists(),
-            "Merged output file {:?} must exist after compaction",
-            merged_path
-        );
-
-        // All input files must be gone (consumed by compaction)
-        for p in &input_paths {
-            assert!(
-                !p.exists(),
-                "Input file {:?} should have been deleted after compaction",
-                p
-            );
-        }
-
-        // maintenance_stats() must reflect the operation
-        let stats = engine.maintenance_stats();
-        assert_eq!(
-            stats.compactions_completed, 1,
-            "compactions_completed must be 1"
-        );
-        assert_eq!(
-            stats.sstables_merged_in, 4,
-            "Should have consumed 4 input SSTables"
-        );
-        assert_eq!(stats.sstables_produced, 1, "sstables_produced must be 1");
-        // bytes_written may be 0 if the merged output is empty (reader/writer compatibility),
-        // but total_time must be non-zero
-        assert!(stats.total_time > Duration::ZERO, "total_time must be > 0");
-    }
-
-    /// #935 branch-review regression: `scan_sstable_candidates` walks the whole
-    /// `data_dir` recursively, so a foreign keyspace/table's SSTable sitting under
-    /// `data_dir` must NOT be treated as a candidate for this table's compaction.
-    /// Before the fix the foreign SSTable inflated `candidate_set`, so a full
-    /// compaction of this table was misclassified as partial (the policy could
-    /// also see the foreign input). After the fix candidates are scoped to
-    /// `data_dir/keyspace/table/`, so only this table's SSTables are merged and
-    /// the foreign file is left untouched.
-    #[test]
-    fn test_maintenance_step_ignores_foreign_table_sstables() {
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let policy = crate::storage::write_engine::STCSPolicy::new(
-            4,   // min_threshold
-            32,  // max_threshold
-            0.5, // bucket_low
-            1.5, // bucket_high
-            0,   // min_sstable_size = 0 so tiny files group together
-        )
-        .unwrap();
-
-        let data_dir = temp_dir.path().join("data");
-        let config = WriteEngineConfig::new(data_dir.clone(), temp_dir.path().join("wal"), schema);
-
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Flush 4 SSTables for THIS table (data/test_ks/test_table/).
-        let input_paths = flush_n_sstables_sync(&mut engine, 4);
-        assert_eq!(input_paths.len(), 4, "Expected 4 flushed SSTables");
-
-        // Plant a foreign keyspace/table SSTable under the same data_dir, with a
-        // sibling TOC.txt so it passes the publication barrier and would be
-        // discovered by the recursive scan.
-        let foreign_dir = data_dir.join("other_ks").join("other_tbl");
-        std::fs::create_dir_all(&foreign_dir).unwrap();
-        let foreign_data = foreign_dir.join("nb-1-big-Data.db");
-        std::fs::write(&foreign_data, b"not a real sstable").unwrap();
-        std::fs::write(foreign_dir.join("nb-1-big-TOC.txt"), b"Data.db\nTOC.txt\n").unwrap();
-
-        engine.set_merge_policy(Box::new(policy)).unwrap();
-        let report = engine.maintenance_step(Duration::from_secs(60)).unwrap();
-
-        // The merge must complete using ONLY this table's 4 inputs.
-        assert_eq!(
-            report.completed_merges.len(),
-            1,
-            "Expected exactly 1 completed merge, got: {:?}",
-            report.completed_merges
-        );
-        let stats = engine.maintenance_stats();
-        assert_eq!(
-            stats.sstables_merged_in, 4,
-            "Only this table's 4 SSTables must be merged; the foreign SSTable must be excluded"
-        );
-
-        // The foreign SSTable must be left completely untouched.
-        assert!(
-            foreign_data.exists(),
-            "Foreign-table SSTable {:?} must not be consumed by this table's compaction",
-            foreign_data
-        );
-
-        // This table's inputs are consumed as usual.
-        for p in &input_paths {
-            assert!(
-                !p.exists(),
-                "Input file {:?} should have been deleted after compaction",
-                p
-            );
-        }
-    }
-
-    #[test]
-    fn test_maintenance_stats_accumulate_across_cycles() {
-        // Run two compaction cycles and verify that stats accumulate.
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let policy = crate::storage::write_engine::STCSPolicy::new(
-            4, 32, 0.5, 1.5, 0, // min_sstable_size=0 for small test files
-        )
-        .unwrap();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-        engine.set_merge_policy(Box::new(policy)).unwrap();
-
-        // First cycle: flush 4, compact
-        flush_n_sstables_sync(&mut engine, 4);
-        engine.maintenance_step(Duration::from_secs(60)).unwrap();
-
-        let stats_after_first = engine.maintenance_stats();
-        assert_eq!(stats_after_first.compactions_completed, 1);
-
-        // Second cycle: flush 4 more, compact again
-        // Row IDs must not collide with the first cycle so each cycle produces 4 SSTables.
-        // flush_n_sstables_sync uses batch * 100 + row, so offset the start batch.
-        // We re-use the helper but note generation counter now starts at a higher value,
-        // so the output SSTable won't conflict with input paths from cycle 1.
-        flush_n_sstables_sync(&mut engine, 4);
-        engine.maintenance_step(Duration::from_secs(60)).unwrap();
-
-        let stats_after_second = engine.maintenance_stats();
-        assert_eq!(
-            stats_after_second.compactions_completed, 2,
-            "Stats must accumulate across compaction cycles"
-        );
-        assert_eq!(
-            stats_after_second.sstables_merged_in, 8,
-            "Should have consumed 8 total input SSTables (2 cycles × 4 each)"
-        );
-        assert_eq!(
-            stats_after_second.sstables_produced, 2,
-            "Should have produced 2 output SSTables"
-        );
-        assert!(
-            stats_after_second.total_time >= stats_after_first.total_time,
-            "Cumulative total_time must only increase"
-        );
-    }
-
-    #[test]
-    fn test_maintenance_step_inputs_intact_on_unwriteable_tmp_dir() {
-        // Failure injection: make the data_dir read-only so creating the tmp
-        // compaction directory fails. All input SSTables must remain intact.
-        //
-        // Note: This test relies on filesystem permissions and is skipped when
-        // running as root (where permissions are not enforced).
-
-        // Skip if running as root (CI containers sometimes run as root)
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::MetadataExt;
-            // Try /proc/self first (Linux), fall back to checking euid via libc
-            let is_root = std::fs::metadata("/proc/self")
-                .map(|m| m.uid() == 0)
-                .unwrap_or_else(|_| {
-                    // On macOS, /proc/self doesn't exist; use a writable sentinel
-                    false
-                });
-            // Also check by trying to write to /etc/cqlite-test-root-check
-            let is_root_macos = std::fs::write("/etc/cqlite-test-root-check", b"")
-                .map(|_| {
-                    let _ = std::fs::remove_file("/etc/cqlite-test-root-check");
-                    true
-                })
-                .unwrap_or(false);
-            if is_root || is_root_macos {
-                // Running as root — permission denial won't work; skip.
-                return;
-            }
-        }
-
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Flush 4 SSTables so STCS can select them
-        let input_paths = flush_n_sstables_sync(&mut engine, 4);
-        for p in &input_paths {
-            assert!(
-                p.exists(),
-                "Input file {:?} should exist before failure test",
-                p
-            );
-        }
-
-        // Make data_dir read-only so creating tmp dir fails
-        let data_dir = temp_dir.path().join("data");
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            std::fs::set_permissions(
-                &data_dir,
-                std::fs::Permissions::from_mode(0o555), // read+execute, no write
-            )
-            .unwrap();
-        }
-
-        let policy = crate::storage::write_engine::STCSPolicy::new(4, 32, 0.5, 1.5, 0).unwrap();
-        engine.set_merge_policy(Box::new(policy)).unwrap();
-
-        // maintenance_step should fail because it cannot create the tmp directory
-        let result = engine.maintenance_step(Duration::from_secs(60));
-
-        // Restore permissions before asserting (so TempDir can clean up)
-        #[cfg(unix)]
-        {
-            use std::os::unix::fs::PermissionsExt;
-            std::fs::set_permissions(&data_dir, std::fs::Permissions::from_mode(0o755)).unwrap();
-        }
-
-        assert!(
-            result.is_err(),
-            "maintenance_step should return an error when the tmp dir cannot be created"
-        );
-
-        // All input files must still exist (atomicity guarantee)
-        for p in &input_paths {
-            assert!(
-                p.exists(),
-                "Input file {:?} must remain intact after failed compaction",
-                p
-            );
-        }
-
-        // Stats must NOT have incremented (no successful compaction)
-        let stats = engine.maintenance_stats();
-        assert_eq!(
-            stats.compactions_completed, 0,
-            "compactions_completed must not increment on failure"
-        );
-    }
-
-    #[test]
-    fn test_no_tmp_dir_remains_after_successful_merge() {
-        // After a successful compaction, the .compaction-tmp-* directory must be cleaned up.
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let policy = crate::storage::write_engine::STCSPolicy::new(4, 32, 0.5, 1.5, 0).unwrap();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-        flush_n_sstables_sync(&mut engine, 4);
-
-        engine.set_merge_policy(Box::new(policy)).unwrap();
-        engine.maintenance_step(Duration::from_secs(60)).unwrap();
-
-        // Scan data_dir for any leftover .compaction-tmp-* directories
-        let data_dir = temp_dir.path().join("data");
-        let leftover_tmp: Vec<_> = std::fs::read_dir(&data_dir)
-            .unwrap()
-            .filter_map(|e| e.ok())
-            .filter(|e| {
-                e.file_name()
-                    .to_string_lossy()
-                    .starts_with(".compaction-tmp-")
-            })
-            .collect();
-
-        assert!(
-            leftover_tmp.is_empty(),
-            "No .compaction-tmp-* directories should remain after successful compaction, \
-             found: {:?}",
-            leftover_tmp.iter().map(|e| e.path()).collect::<Vec<_>>()
-        );
-    }
-
-    // ============================================================================
-    // Issue #474 review follow-up: NB-1 startup sweep tests
-    // ============================================================================
-
-    /// Helper: build a WriteEngineConfig pointing at a given data/wal dir pair.
-    fn config_for(temp_dir: &TempDir) -> WriteEngineConfig {
-        WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            create_test_schema(),
-        )
-    }
-
-    #[test]
-    fn test_startup_sweep_removes_orphaned_compaction_tmp_dir() {
-        // Pre-seed a .compaction-tmp-99/ directory under data_dir.
-        // WriteEngine::new() must remove it during the startup sweep.
-        let temp_dir = TempDir::new().unwrap();
-        let data_dir = temp_dir.path().join("data");
-        std::fs::create_dir_all(&data_dir).unwrap();
-
-        let orphan_dir = data_dir.join(".compaction-tmp-99");
-        std::fs::create_dir_all(&orphan_dir).unwrap();
-        // Put a partial component file inside to make it non-trivially non-empty
-        std::fs::write(orphan_dir.join("partial.db"), b"partial content").unwrap();
-
-        assert!(
-            orphan_dir.exists(),
-            "Orphan dir should exist before engine creation"
-        );
-
-        let config = config_for(&temp_dir);
-        let _engine = WriteEngine::new(config).unwrap();
-
-        assert!(
-            !orphan_dir.exists(),
-            "Startup sweep must remove orphaned .compaction-tmp-99/ directory"
-        );
-    }
-
-    #[test]
-    fn test_startup_sweep_removes_orphaned_partial_sstable() {
-        // Pre-seed nb-99-big-Data.db (and friends) WITHOUT a TOC.txt in the
-        // keyspace/table subdirectory.  WriteEngine::new() must remove them.
-        let temp_dir = TempDir::new().unwrap();
-        let data_dir = temp_dir.path().join("data");
-        let sstable_dir = data_dir.join("test_ks").join("test_table");
-        std::fs::create_dir_all(&sstable_dir).unwrap();
-
-        // Create orphaned components (no TOC.txt)
-        let orphan_components = [
-            "nb-99-big-Data.db",
-            "nb-99-big-Index.db",
-            "nb-99-big-Statistics.db",
-        ];
-        for name in &orphan_components {
-            std::fs::write(sstable_dir.join(name), b"orphaned").unwrap();
-        }
-
-        // Also create a complete SSTable (has TOC.txt) — must NOT be touched
-        let complete_components = ["nb-1-big-Data.db", "nb-1-big-Index.db", "nb-1-big-TOC.txt"];
-        for name in &complete_components {
-            std::fs::write(sstable_dir.join(name), b"good").unwrap();
-        }
-
-        let config = config_for(&temp_dir);
-        let _engine = WriteEngine::new(config).unwrap();
-
-        // Orphaned components must be gone
-        for name in &orphan_components {
-            assert!(
-                !sstable_dir.join(name).exists(),
-                "Startup sweep must remove orphaned component {:?}",
-                name
-            );
-        }
-
-        // Complete SSTable must remain intact
-        for name in &complete_components {
-            assert!(
-                sstable_dir.join(name).exists(),
-                "Startup sweep must NOT remove complete SSTable component {:?}",
-                name
-            );
-        }
-    }
-
-    #[test]
-    fn test_startup_sweep_leaves_complete_sstable_intact() {
-        // A full SSTable set (Data.db + TOC.txt + others) must survive the sweep.
-        let temp_dir = TempDir::new().unwrap();
-        let data_dir = temp_dir.path().join("data");
-        let sstable_dir = data_dir.join("test_ks").join("test_table");
-        std::fs::create_dir_all(&sstable_dir).unwrap();
-
-        let all_components = [
-            "nb-3-big-Data.db",
-            "nb-3-big-Index.db",
-            "nb-3-big-Summary.db",
-            "nb-3-big-Statistics.db",
-            "nb-3-big-Filter.db",
-            "nb-3-big-Digest.crc32",
-            "nb-3-big-TOC.txt",
-        ];
-        for name in &all_components {
-            std::fs::write(sstable_dir.join(name), b"complete").unwrap();
-        }
-
-        let config = config_for(&temp_dir);
-        let _engine = WriteEngine::new(config).unwrap();
-
-        for name in &all_components {
-            assert!(
-                sstable_dir.join(name).exists(),
-                "Complete SSTable component {:?} must not be removed by startup sweep",
-                name
-            );
-        }
-    }
-
-    #[test]
-    fn test_bytes_written_includes_all_components() {
-        // After a successful merge, cumulative_stats.bytes_written must be at
-        // least as large as the sum of Data.db sizes alone (i.e. it includes
-        // the other component files too).
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-
-        let policy = crate::storage::write_engine::STCSPolicy::new(4, 32, 0.5, 1.5, 0).unwrap();
-
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-
-        let mut engine = WriteEngine::new(config).unwrap();
-        let input_paths = flush_n_sstables_sync(&mut engine, 4);
-
-        // Compute the sum of just the Data.db sizes before compaction
-        let data_db_total: u64 = input_paths
-            .iter()
-            .map(|p| std::fs::metadata(p).map(|m| m.len()).unwrap_or(0))
-            .sum();
-
-        engine.set_merge_policy(Box::new(policy)).unwrap();
-        engine.maintenance_step(Duration::from_secs(60)).unwrap();
-
-        let stats = engine.maintenance_stats();
-        // bytes_written counts all output components, so it should be >= what
-        // data_db_total reported for the inputs (output may differ in size but
-        // the multi-component sum must be >= the Data.db-only measurement).
-        // More concretely: if any non-Data component was written, the total
-        // must be larger than data_size alone.
-        //
-        // We assert >= 0 always holds (u64), and additionally that the field
-        // was updated at all (compaction ran).
-        assert_eq!(stats.compactions_completed, 1, "compaction must have run");
-        // The bytes_written field is now the sum of all components.
-        // We can't assert an exact value, but we know:
-        //  - data_db_total may be 0 for tiny test SSTables written by the test writer
-        //  - if data_db_total > 0, bytes_written >= data_db_total is a reasonable lower bound
-        //  - at minimum, the field must equal total_bytes_written (multi-component sum) >= 0
-        let _ = data_db_total; // used above for context; value may be 0 in test environment
-                               // The assertion that matters: stats are populated and consistent across calls.
-                               // maintenance_stats() returns a clone so two consecutive calls must agree.
-        let stats2 = engine.maintenance_stats();
-        assert_eq!(
-            stats.bytes_written, stats2.bytes_written,
-            "maintenance_stats() must be consistent across calls"
-        );
-        // bytes_written is u64; it is always >= 0. Just confirm the field was set.
-        assert_eq!(
-            stats.sstables_produced, 1,
-            "one output SSTable must have been produced"
-        );
-    }
-
     // ============================================================================
     // Issue #547: WAL durability toggle tests
     // ============================================================================
@@ -4239,157 +2065,6 @@ mod tests {
             engine2.is_ok(),
             "WriteEngine must acquire lock after the previous engine was dropped: {:?}",
             engine2.err()
-        );
-    }
-
-    // -----------------------------------------------------------------------
-    // Issue #486 — total_written and l0_count non-placeholder behaviour
-    // -----------------------------------------------------------------------
-
-    /// After writing N rows and flushing, `total_written()` == N even though
-    /// `memtable_row_count()` has been reset to 0.
-    #[tokio::test]
-    async fn test_total_written_survives_flush() {
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Write 5 rows to the first "batch"
-        for i in 0..5 {
-            engine
-                .write(create_test_mutation(
-                    i,
-                    &format!("User{i}"),
-                    1_000_000 + i as i64,
-                ))
-                .unwrap();
-        }
-        assert_eq!(engine.total_written(), 5);
-        assert_eq!(engine.memtable_row_count(), 5);
-
-        // Flush — memtable resets but total_written must NOT
-        engine.flush().await.unwrap();
-        assert_eq!(
-            engine.memtable_row_count(),
-            0,
-            "memtable should be empty after flush"
-        );
-        assert_eq!(
-            engine.total_written(),
-            5,
-            "total_written must NOT reset after flush"
-        );
-
-        // Write 3 more rows in a second batch
-        for i in 10..13 {
-            engine
-                .write(create_test_mutation(
-                    i,
-                    &format!("User{i}"),
-                    2_000_000 + i as i64,
-                ))
-                .unwrap();
-        }
-        assert_eq!(
-            engine.total_written(),
-            8,
-            "total_written must accumulate across flushes"
-        );
-        assert_eq!(engine.memtable_row_count(), 3);
-
-        // Flush again
-        engine.flush().await.unwrap();
-        assert_eq!(engine.memtable_row_count(), 0);
-        assert_eq!(
-            engine.total_written(),
-            8,
-            "total_written must still be 8 after second flush"
-        );
-    }
-
-    /// `l0_count()` reflects the number of successful flush operations (not zero).
-    #[tokio::test]
-    async fn test_l0_count_increments_on_flush() {
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        assert_eq!(engine.l0_count(), 0, "l0_count should start at 0");
-
-        // Flushing an empty memtable produces no SSTable — count stays 0
-        engine.flush().await.unwrap();
-        assert_eq!(
-            engine.l0_count(),
-            0,
-            "empty flush must not increment l0_count"
-        );
-
-        // Write one row and flush — count must become 1
-        engine
-            .write(create_test_mutation(1, "Alice", 1_000_000))
-            .unwrap();
-        engine.flush().await.unwrap();
-        assert_eq!(
-            engine.l0_count(),
-            1,
-            "l0_count must be 1 after first non-empty flush"
-        );
-
-        // Write and flush again — count must become 2
-        engine
-            .write(create_test_mutation(2, "Bob", 2_000_000))
-            .unwrap();
-        engine.flush().await.unwrap();
-        assert_eq!(
-            engine.l0_count(),
-            2,
-            "l0_count must be 2 after second non-empty flush"
-        );
-    }
-
-    /// Verify that `total_written` != `memtable_row_count` after a flush —
-    /// the scenario that proves the placeholder is replaced.
-    #[tokio::test]
-    async fn test_total_written_differs_from_memtable_after_flush() {
-        let temp_dir = TempDir::new().unwrap();
-        let schema = create_test_schema();
-        let config = WriteEngineConfig::new(
-            temp_dir.path().join("data"),
-            temp_dir.path().join("wal"),
-            schema,
-        );
-        let mut engine = WriteEngine::new(config).unwrap();
-
-        // Write 7 rows and flush
-        for i in 0..7 {
-            engine
-                .write(create_test_mutation(
-                    i,
-                    &format!("User{i}"),
-                    1_000_000 + i as i64,
-                ))
-                .unwrap();
-        }
-        engine.flush().await.unwrap();
-
-        // Post-flush: memtable_row_count is 0 but total_written is 7
-        assert_eq!(engine.memtable_row_count(), 0);
-        assert_eq!(engine.total_written(), 7);
-        assert_ne!(
-            engine.total_written() as usize,
-            engine.memtable_row_count(),
-            "total_written must differ from memtable_row_count after flush — \
-             proves neither is a placeholder"
         );
     }
 }
